Move octomap stuff to the spacetime directory
