+++ opencog/spacetime/CMakeLists.txt
@@ -1,2 +1,5 @@
 
 ADD_SUBDIRECTORY (atom-types)
+IF (HAVE_OCTOMAP)
+	ADD_SUBDIRECTORY (octomap)
+ENDIF (HAVE_OCTOMAP)
