+++ opencog/spacetime/octomap/AtomOcTree.cc
