+++ opencog/spacetime/octomap/AtomOcTree.h
