+++ opencog/spacetime/octomap/AtomOcTreeNode.cc
