+++ opencog/spacetime/octomap/AtomOcTreeNode.h
