+++ opencog/spacetime/octomap/CMakeLists.txt
@@ -1,15 +1,15 @@
-ADD_LIBRARY (time_octomap SHARED
+ADD_LIBRARY (octomap SHARED
 	AtomOcTree.cc
 	AtomOcTreeNode.cc
 	TimeOctomap.cc
 )
 
-TARGET_LINK_LIBRARIES(time_octomap
+TARGET_LINK_LIBRARIES(octomap
 	${OCTOMAP_LIBRARY}
 	${OCTOMAP_OCTOMATH_LIBRARY}
 )
 
-INSTALL (TARGETS time_octomap
+INSTALL (TARGETS octomap
 	DESTINATION "lib${LIB_DIR_SUFFIX}/opencog"
 )
 
@@ -17,7 +17,7 @@ INSTALL (FILES
 	AtomOcTree.h
 	AtomOcTreeNode.h
 	TimeOctomap.h
-	DESTINATION "include/${PROJECT_NAME}/timeoctomap"
+	DESTINATION "include/${PROJECT_NAME}/spacetime/octomap"
 )
 
 ADD_SUBDIRECTORY (pointmemory)
