+++ opencog/spacetime/octomap/pointmemory/pointmem.scm
